Updated the Chart.yaml to use the basic symbols
tried to standardize the code formatting to keep it neat to. I know I'm adding a lot of lines.

I also found a way to make the stitch legend mutable so I can just use the methods already there and replace with my stitch legend instead. Kudos to the previous author for making that an option